Comment에서 Get요청시 불필요한 로그인 확인 제거, 쿼리문 수정(@Param값 삽입), 기능 구현 완료로 인하여 불필요해진 sout제거

+++ src/main/java/com/hanghae/minipj/Service/CommentService.java
@@ -46,7 +46,14 @@ public class CommentService {
         }
         Comment comment = new Comment(requestDto.getContent(), card, member);
         commentRepository.save(comment);
-        return ResponseDto.success(null);
+        CommentResponseDto commentResponseDto = CommentResponseDto.builder()
+                .id(comment.getId())
+                .nickname(comment.getMember().getNickname())
+                .content(comment.getContent())
+                .createdAt(comment.getCreatedAt())
+                .isEditMode(comment.getIsEditMode())
+                .build();
+        return ResponseDto.success(commentResponseDto);
     }
 
     public ResponseDto<?> getComments() {
@@ -83,7 +90,7 @@ public class CommentService {
         }
 
         if (comment.validateMember(member)) {
-            return ResponseDto.fail("BAD_REQUEST", "작성자만 수정할 수 있습니다.");
+            return ResponseDto.fail("UNAUTHORIZED", "작성자만 수정할 수 있습니다.");
         }
         comment.update(requestDto);
         CommentResponseDto commentResponseDto= CommentResponseDto.builder()
@@ -106,7 +113,11 @@ public class CommentService {
         if (null == member) {
             return ResponseDto.fail("INVALID_TOKEN", "Token이 유효하지 않습니다.");
         }
+
        Comment comment =isPresentComment(id);
+        if (comment.validateMember(member)) {
+            return ResponseDto.fail("UNAUTHORIZED", "작성자만 삭제할 수 있습니다.");
+        }
         commentRepository.delete(comment);
         return ResponseDto.success(null);
     }
@@ -130,15 +141,7 @@ public class CommentService {
         return tokenProvider.getMemberFromAuthentication();
     }
 
-    public ResponseDto<?> getCommentByCardId(HttpServletRequest request, Long id) {
-        if (null == request.getHeader("Authorization")) {
-            return ResponseDto.fail("MEMBER_NOT_FOUND",
-                    "로그인이 필요합니다.");
-        }
-        Member member = validateMember(request);
-        if (null == member) {
-            return ResponseDto.fail("INVALID_TOKEN", "Token이 유효하지 않습니다.");
-        }
+    public ResponseDto<?> getCommentByCardId(Long id) {
         List<Comment> commentList =commentRepository.findAllByCardIdOrderByCreatedAtDesc(id);
         List<CommentResponseDto> commentResponseDtoList =new ArrayList<>();
         for(Comment comment:commentList){
@@ -148,6 +151,7 @@ public class CommentService {
                             .nickname(comment.getMember().getNickname())
                             .content(comment.getContent())
                             .createdAt(comment.getCreatedAt())
+                            .isEditMode(comment.getIsEditMode())
                             .build()
             );
         }
